+++ cmd/drone-server/inject_runner.go
@@ -38,9 +38,9 @@ func provideRunner(
 	registry core.RegistryService,
 	config config.Config,
 ) *runner.Runner {
-	// the local runner is only created when the nomad scheduler,
-	// kubernetes scheduler, and remote agents are disabled
-	if config.Nomad.Enabled || config.Kube.Enabled || (config.Agent.Disabled == false) {
+	// the local runner is only created when remote agents
+	// are disabled
+	if config.Agent.Disabled == false {
 		return nil
 	}
 	engine, err := docker.NewEnv()