+++ metric/sink/config.go
@@ -19,17 +19,15 @@ type Config struct {
 	Endpoint string
 	Token    string
 
-	License          string
-	Licensor         string
-	Subscription     string
-	EnableGithub     bool
-	EnableGithubEnt  bool
-	EnableGitlab     bool
-	EnableBitbucket  bool
-	EnableStash      bool
-	EnableGogs       bool
-	EnableGitea      bool
-	EnableAgents     bool
-	EnableNomad      bool
-	EnableKubernetes bool
+	License         string
+	Licensor        string
+	Subscription    string
+	EnableGithub    bool
+	EnableGithubEnt bool
+	EnableGitlab    bool
+	EnableBitbucket bool
+	EnableStash     bool
+	EnableGogs      bool
+	EnableGitea     bool
+	EnableAgents    bool
 }