+++ metric/sink/tags.go
@@ -47,10 +47,6 @@ func createTags(config Config) []string {
 	switch {
 	case config.EnableAgents:
 		tags = append(tags, "scheduler:internal:agents")
-	case config.EnableKubernetes:
-		tags = append(tags, "scheduler:kubernetes")
-	case config.EnableNomad:
-		tags = append(tags, "scheduler:nomad")
 	default:
 		tags = append(tags, "scheduler:internal:local")
 	}