+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Breaking
+- removed deprecated kubernetes integration in favor of official kubernetes runner.
+- removed deprecated nomad integration in favor of official nomad runner.
+
 ## [1.10.1]
 ### Added
 - support for repository-level concurrency limits.